+++ src/modules/students/student.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { StudentRepository } from './student.repository';
 import { CreateStudentDto } from './dto/createStudent.dto';
 import { Student } from 'src/modules/students/entites/student.entity';
+import { ItemDoesntExistException } from 'src/exceptions/ItemDoesntExist.exception';
 
 @Injectable()
 export class StudentService {
@@ -12,14 +13,32 @@ export class StudentService {
   }
 
   async deleteStudent(studentId: string): Promise<void> {
+    const student = await this.studentRepository.findStudentById(studentId);
+
+    if (!student) {
+      throw new ItemDoesntExistException('Student', studentId);
+    }
+
     await this.studentRepository.deleteStudent(studentId);
   }
 
   async unassignStudentFronClass(studentId: string): Promise<void> {
+    const student = await this.studentRepository.findStudentById(studentId);
+
+    if (!student) {
+      throw new ItemDoesntExistException('Student', studentId);
+    }
+
     await this.studentRepository.unassignStudentFronClass(studentId);
   }
 
   async assignClass(studentId: string, classId: number): Promise<void> {
+     const student = await this.studentRepository.findStudentById(studentId);
+
+    if (!student) {
+      throw new ItemDoesntExistException('Student', studentId)
+    }
+    
     await this.studentRepository.assignClass(studentId, classId);
   }
